deribit_client/pyproject.toml
[tool.poetry]
name = "deribit-client"
version = "0.1.0"
description = ""
authors = [" leonideliseev <leonid.2004eliseev@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.36"
aiosqlite = "^0.20.0"
aiohttp = "^3.10.10"
asyncpg = "^0.30.0"
asyncio = "^3.4.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

